Adding delay before running clients
On some instances clients may be too quick - although the client container
depends on the server being up and running, that doesn't mean the service is
actually ready to serve requests. So, we give it a few seconds before trying. In
reality this will add about a minute at worst to a process that already takes
about 10 minutes, which isn't great but at least certain failures should stop
now, I guess